factor out annotateEvent and isAnnotated from lib/carelink/common
