+++ lib/carelink/bolusNWizard.js
@@ -17,6 +17,7 @@
 
 var _ = require('lodash');
 
+var annotate = require('../eventAnnotations');
 var common = require('./common.js');
 var parsing = require('./parsing.js');
 
@@ -176,7 +177,7 @@ module.exports = function(timezone, model) {
               return null;
             }
             if (i !== 0) {
-              obj = common.annotateEvent(obj, 'carelink/wizard/long-search');
+              obj = annotate.annotateEvent(obj, 'carelink/wizard/long-search');
             }
             return obj;
           }
@@ -222,7 +223,7 @@ module.exports = function(timezone, model) {
               obj = datumsBySeqNum[parsed.uploadId][parsed.uploadSeqNum - (3 + i)];
               obj = obj ? parser(obj) : null;
               if (obj != null) {
-                obj = common.annotateEvent(obj, 'carelink/wizard/long-search');
+                obj = annotate.annotateEvent(obj, 'carelink/wizard/long-search');
               }
             }
           }
@@ -232,7 +233,7 @@ module.exports = function(timezone, model) {
               return null;
             }
             if (i !== 0) {
-              obj = common.annotateEvent(obj, 'carelink/wizard/long-search');
+              obj = annotate.annotateEvent(obj, 'carelink/wizard/long-search');
             }
             return obj;
           }
@@ -285,7 +286,7 @@ module.exports = function(timezone, model) {
               return null;
             }
             if (i !== 0) {
-              obj = common.annotateEvent(obj, 'carelink/wizard/long-search');
+              obj = annotate.annotateEvent(obj, 'carelink/wizard/long-search');
             }
             return obj;
           }
@@ -312,7 +313,7 @@ module.exports = function(timezone, model) {
           else {
             obj = datumsBySeqNum[parsed.uploadId][parsed.uploadSeqNum + 2];
             obj = obj ? parser(obj) : null;
-            obj = obj ? common.annotateEvent(obj, 'carelink/wizard/long-search') : null;
+            obj = obj ? annotate.annotateEvent(obj, 'carelink/wizard/long-search') : null;
           }
           if (obj != null && obj.type === 'bolusSquare') {
             return obj;
@@ -332,7 +333,7 @@ module.exports = function(timezone, model) {
               return null;
             }
             if (i !== 0) {
-              obj = common.annotateEvent(obj, 'carelink/wizard/long-search');
+              obj = annotate.annotateEvent(obj, 'carelink/wizard/long-search');
             }
             return obj;
           }
@@ -368,7 +369,7 @@ module.exports = function(timezone, model) {
           if (obj == null) {
             obj = datumsBySeqNum[bolus.uploadId][bolus.uploadSeqNum + 2];
             obj = obj ? parser(obj) : null;
-            obj = obj ? common.annotateEvent(obj, 'carelink/wizard/long-search') : null;
+            obj = obj ? annotate.annotateEvent(obj, 'carelink/wizard/long-search') : null;
           }
           if (obj != null && obj.type === 'wizard') {
             return obj;
@@ -405,7 +406,7 @@ module.exports = function(timezone, model) {
             if (obj == null) {
               obj = datumsBySeqNum[bolus.uploadId][bolus.uploadSeqNum + 3];
               obj = obj ? parser(obj) : null;
-              obj = obj ? common.annotateEvent(obj, 'carelink/wizard/long-search') : null;
+              obj = obj ? annotate.annotateEvent(obj, 'carelink/wizard/long-search') : null;
             }
             if (obj != null && obj.type === 'wizard') {
               return obj;
@@ -420,7 +421,7 @@ module.exports = function(timezone, model) {
             if (obj == null) {
               obj = datumsBySeqNum[bolus.uploadId][bolus.uploadSeqNum + 2];
               obj = obj ? parser(obj) : null;
-              obj = obj ? common.annotateEvent(obj, 'carelink/wizard/long-search') : null;
+              obj = obj ? annotate.annotateEvent(obj, 'carelink/wizard/long-search') : null;
             }
             if (obj != null && obj.type === 'wizard') {
               return obj;
