+++ lib/carelink/settings.js
@@ -17,6 +17,7 @@
 
 var _ = require('lodash');
 
+var annotate = require('../eventAnnotations');
 var common = require('./common.js');
 var parsing = require('./parsing.js');
 
@@ -241,7 +242,7 @@ module.exports = function (timezone) {
             delete prevSettings.annotations;
 
             if (datum['Raw-Values']['PATTERN_NAME'] !== currSettings.activeSchedule) {
-              common.annotateEvent(prevSettings, 'carelink/settings/activeSchedule-mismatch');
+              annotate.annotateEvent(prevSettings, 'carelink/settings/activeSchedule-mismatch');
             }
 
 
@@ -289,7 +290,7 @@ module.exports = function (timezone) {
             delete prevSettings.annotations;
 
             if (!_.isEqual(newSchedule, currSettings.basalSchedules[scheduleName])) {
-              common.annotateEvent(prevSettings, 'carelink/settings/basal-mismatch');
+              annotate.annotateEvent(prevSettings, 'carelink/settings/basal-mismatch');
             }
 
 
@@ -326,7 +327,7 @@ module.exports = function (timezone) {
             delete prevSettings.annotations;
 
             if (!_.isEqual(newConfig, _.pick(currSettings, Object.keys(newConfig)))) {
-              common.annotateEvent(prevSettings, 'carelink/settings/wizard-mismatch');
+              annotate.annotateEvent(prevSettings, 'carelink/settings/wizard-mismatch');
             }
 
             if (common.isSuspectedNewDevice(prevSettings)) {
